requirements.txt
pytest>=7.1.2
selene>=2.0.0b8
allure-pytest




